+++ test/pytest_cases/test_pipetask.py
@@ -47,18 +47,30 @@ class TestPipeTask(object):
         self.pipeline_db_folder = os.path.join(self.top_path, "results", "pipeline_info")
         self.pipenode_db_folder = os.path.join(self.top_path, "results", "pipenode_info")
         self.pipetask_db_folder = os.path.join(self.top_path, "results", "pipetask_info")
-        all_file_name_list = os.listdir(self.pipenode_db_folder)
-        all_test_ahead_list = [i for i in all_file_name_list if i.startswith("test_")]
-        for test_file in all_test_ahead_list:
-            os.remove(os.path.join(self.pipenode_db_folder, test_file))
-        all_file_name_list = os.listdir(self.pipeline_db_folder)
-        all_test_ahead_list = [i for i in all_file_name_list if i.startswith("test_")]
-        for test_file in all_test_ahead_list:
-            os.remove(os.path.join(self.pipeline_db_folder, test_file))
-        all_file_name_list = os.listdir(self.pipetask_db_folder)
-        all_test_ahead_list = [i for i in all_file_name_list if i.startswith("test_")]
-        for test_file in all_test_ahead_list:
-            os.remove(os.path.join(self.pipetask_db_folder, test_file))
+
+        if os.path.exists(self.pipenode_db_folder):
+            all_file_name_list = os.listdir(self.pipenode_db_folder)
+            all_test_ahead_list = [i for i in all_file_name_list if i.startswith("test_")]
+            for test_file in all_test_ahead_list:
+                os.remove(os.path.join(self.pipenode_db_folder, test_file))
+        else:
+            os.makedirs(self.pipenode_db_folder)
+
+        if os.path.exists(self.pipeline_db_folder):
+            all_file_name_list = os.listdir(self.pipeline_db_folder)
+            all_test_ahead_list = [i for i in all_file_name_list if i.startswith("test_")]
+            for test_file in all_test_ahead_list:
+                os.remove(os.path.join(self.pipeline_db_folder, test_file))
+        else:
+            os.makedirs(self.pipeline_db_folder)
+
+        if os.path.exists(self.pipetask_db_folder):
+            all_file_name_list = os.listdir(self.pipetask_db_folder)
+            all_test_ahead_list = [i for i in all_file_name_list if i.startswith("test_")]
+            for test_file in all_test_ahead_list:
+                os.remove(os.path.join(self.pipetask_db_folder, test_file))
+        else:
+            os.makedirs(self.pipetask_db_folder)
         # 再粘
         self.fake_pipenode_pkl_folder = os.path.join(self.fake_path, "test_results", "fake_pipenode_info")
         self.fake_pipeline_pkl_folder = os.path.join(self.fake_path, "test_results", "fake_pipeline_info")
@@ -102,17 +114,17 @@ class TestPipeTask(object):
         for appeared_id in self.appeared_ppn_id_list:
             file_path = os.path.join(self.pipenode_db_folder, appeared_id + ".pkl")
             if os.path.exists(file_path):
-                logger.warning("find ppn_id={} still exists, remove it".format(appeared_id))
+                # logger.warning("find ppn_id={} still exists, remove it".format(appeared_id))
                 os.remove(file_path)
         for appeared_id in self.appeared_ppl_id_list:
             file_path = os.path.join(self.pipeline_db_folder, appeared_id + ".pkl")
             if os.path.exists(file_path):
-                logger.warning("find ppl_id={} still exists, remove it".format(appeared_id))
+                # logger.warning("find ppl_id={} still exists, remove it".format(appeared_id))
                 os.remove(file_path)
         for appeared_id in self.appeared_ppt_id_list:
             file_path = os.path.join(self.pipetask_db_folder, appeared_id + ".pkl")
             if os.path.exists(file_path):
-                logger.warning("find ppt_id={} still exists, remove it".format(appeared_id))
+                # logger.warning("find ppt_id={} still exists, remove it".format(appeared_id))
                 os.remove(file_path)
 
     def test_pipetask_create(self):
@@ -268,17 +280,17 @@ class TestSpecialWorkflow(object):
         for appeared_id in self.appeared_ppn_id_list:
             file_path = os.path.join(self.pipenode_db_folder, appeared_id + ".pkl")
             if os.path.exists(file_path):
-                logger.warning("find ppn_id={} still exists, remove it".format(appeared_id))
+                # logger.warning("find ppn_id={} still exists, remove it".format(appeared_id))
                 os.remove(file_path)
         for appeared_id in self.appeared_ppl_id_list:
             file_path = os.path.join(self.pipeline_db_folder, appeared_id + ".pkl")
             if os.path.exists(file_path):
-                logger.warning("find ppl_id={} still exists, remove it".format(appeared_id))
+                # logger.warning("find ppl_id={} still exists, remove it".format(appeared_id))
                 os.remove(file_path)
         for appeared_id in self.appeared_ppt_id_list:
             file_path = os.path.join(self.pipetask_db_folder, appeared_id + ".pkl")
             if os.path.exists(file_path):
-                logger.warning("find ppt_id={} still exists, remove it".format(appeared_id))
+                # logger.warning("find ppt_id={} still exists, remove it".format(appeared_id))
                 os.remove(file_path)
 
     def test_output_str_func(self):
